Move root core.lib stragglers into block, platform, and behavior packages
