+++ common/src/main/java/com/logistics/automation/laserquarry/LaserQuarryBlock.java
@@ -4,7 +4,7 @@ import com.logistics.automation.laserquarry.entity.LaserQuarryBlockEntity;
 import com.logistics.core.LogisticsConfig;
 import com.logistics.LogisticsAutomation;
 import com.logistics.core.lib.block.behavior.ProbeBehavior;
-import com.logistics.core.lib.support.ProbeResult;
+import com.logistics.core.lib.block.behavior.ProbeResult;
 import com.logistics.automation.marker.MarkerManager;
 import com.mojang.serialization.MapCodec;
 import net.minecraft.network.chat.Component;
