+++ common/src/main/java/com/logistics/automation/marker/MarkerBlockEntity.java
@@ -1,7 +1,7 @@
 package com.logistics.automation.marker;
 
 import com.logistics.LogisticsAutomation;
-import com.logistics.core.lib.BaseBlockEntity;
+import com.logistics.core.lib.block.BaseBlockEntity;
 import com.logistics.core.lib.compat.NbtCompat;
 import java.util.ArrayList;
 import java.util.List;
