+++ common/src/main/java/com/logistics/core/lib/block/BaseBlockEntity.java
@@ -1,4 +1,4 @@
-package com.logistics.core.lib;
+package com.logistics.core.lib.block;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.HolderLookup;
