+++ common/src/main/java/com/logistics/core/lib/block/BlockEntitySupplier.java
@@ -1,4 +1,4 @@
-package com.logistics.core.lib;
+package com.logistics.core.lib.block;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.block.entity.BlockEntity;
