+++ fabric/src/main/resources/META-INF/services/com.logistics.core.lib.block.BlockEntityTypeFactory
@@ -1,0 +1,1 @@
+com.logistics.fabric.FabricBlockEntityBuilder
